+++ src/main/kotlin/com/teamsparta/tikitaka/domain/match/controller/v1/MatchController.kt
@@ -5,6 +5,10 @@ import com.teamsparta.tikitaka.domain.match.dto.MatchStatusResponse
 import com.teamsparta.tikitaka.domain.match.dto.PostMatchRequest
 import com.teamsparta.tikitaka.domain.match.dto.UpdateMatchRequest
 import com.teamsparta.tikitaka.domain.match.service.v1.MatchService
+import org.springframework.data.domain.Page
+import org.springframework.data.domain.Pageable
+import org.springframework.data.domain.Sort
+import org.springframework.data.web.PageableDefault
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
@@ -15,7 +19,7 @@ class MatchController(
     private val matchService: MatchService,
 ) {
     //@PreAuthorize("hasRole('LEADER')") //todo : 리더 외 권한 부여 ?
-    @PostMapping()
+    @PostMapping("/create")
     fun postMatch(@RequestBody request: PostMatchRequest): ResponseEntity<MatchStatusResponse> {
         return ResponseEntity.status(HttpStatus.CREATED).body(matchService.postMatch(request))
     }
@@ -36,9 +40,10 @@ class MatchController(
     }
 
     @GetMapping()
-    fun getMatches(): ResponseEntity<List<MatchResponse>> {
-        return ResponseEntity.status(HttpStatus.OK)
-            .body(matchService.getMatches())
+    fun getMatches(
+        @PageableDefault(size = 10, sort = ["createdAt,desc"]) pageable: Pageable
+    ): ResponseEntity<Page<MatchResponse>> {
+        return ResponseEntity.status(HttpStatus.OK).body(matchService.getMatches(pageable))
     }
 
     @GetMapping("/{match-id}")
