+++ src/main/kotlin/com/teamsparta/tikitaka/domain/match/dto/MatchResponse.kt
@@ -5,17 +5,18 @@ import java.time.LocalDateTime
 
 data class MatchResponse(
 
-    val id: Long,
-    val teamId: Long,
-    val title: String,
-    val matchDate: LocalDateTime,
-    val location: String,
-    val content: String,
-    val matchStatus: Boolean,
+    val id :Long,
+    val teamId : Long,
+    val title:String,
+    val matchDate :LocalDateTime,
+    val location : String,
+    val content : String,
+    val matchStatus : Boolean,
 
-    ) {
-    companion object {
-        fun from(match: Match) = MatchResponse(
+    )
+{
+    companion object{
+        fun from(match : Match) = MatchResponse(
             id = match.id!!,
             teamId = match.teamId,
             title = match.title,
@@ -24,6 +25,6 @@ data class MatchResponse(
             content = match.content,
             matchStatus = match.matchStatus,
 
-            )
+        )
     }
 }
