+++ src/main/kotlin/com/teamsparta/tikitaka/domain/match/dto/MatchStatusResponse.kt
@@ -1,9 +1,9 @@
 package com.teamsparta.tikitaka.domain.match.dto
 
 data class MatchStatusResponse(
-    val status: Boolean,
-) {
-    companion object {
+    val status : Boolean,
+){
+    companion object{
         fun from() = MatchStatusResponse(
             status = true,
         )
