+++ src/main/kotlin/com/teamsparta/tikitaka/domain/match/dto/UpdateMatchRequest.kt
@@ -2,10 +2,10 @@ package com.teamsparta.tikitaka.domain.match.dto
 
 import java.time.LocalDateTime
 
-class UpdateMatchRequest(
-    val title: String,
-    val matchDate: LocalDateTime,
-    val location: String,
-    val content: String,
-    val matchStatus: Boolean,
+class UpdateMatchRequest (
+    val title:String,
+    val matchDate : LocalDateTime,
+    val location : String,
+    val content : String,
+    val matchStatus : Boolean,
 )
