+++ src/main/kotlin/com/teamsparta/tikitaka/domain/match/model/Match.kt
@@ -3,12 +3,12 @@ package com.teamsparta.tikitaka.domain.match.model
 import com.teamsparta.tikitaka.domain.common.baseentity.BaseEntity
 import com.teamsparta.tikitaka.domain.match.dto.UpdateMatchRequest
 import jakarta.persistence.*
+import java.lang.RuntimeException
 import java.time.LocalDateTime
 
 @Entity
 @Table(name = "match_post")
 class Match(
-
     @Column(name = "title", nullable = false)
     var title: String,
 
@@ -26,34 +26,32 @@ class Match(
 
     @Column(name = "post_team_id", nullable = false)
     var teamId: Long
-) : BaseEntity() {
+) : BaseEntity()
+{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     var id: Long? = null
 
     fun updateMatch(
-        request: UpdateMatchRequest
-    ) {
+        request : UpdateMatchRequest
+    )
+    {
         validateTitle(request.title)
         validateMatchDate(request.matchDate)
-        this.title = request.title
-        this.matchDate = request.matchDate
+        this.title=request.title
+        this.matchDate =request.matchDate
         this.location = request.location
         this.content = request.content
         this.matchStatus = request.matchStatus
     }
 
 
-    companion object {
+    companion object{
 
         fun of(
-            matchDate: LocalDateTime,
-            location: String,
-            content: String,
-            matchStatus: Boolean,
-            teamId: Long,
-            title: String
-        ): Match {
+            matchDate: LocalDateTime, location: String, content: String, matchStatus: Boolean, teamId: Long, title: String
+        ): Match
+        {
             return Match(
                 matchDate = matchDate,
                 location = location,
@@ -70,25 +68,28 @@ class Match(
     }
 
     private fun validateTitle(
-        title: String
-    ) {
-        if (title.length > 20) {
+        title:String
+    )
+    {
+        if(title.length >20){
             throw RuntimeException("제목이 너무 깁니다.") //todo : custom exception
         }
     }
 
     private fun validateMatchDate(
         matchDate: LocalDateTime
-    ) {
+    )
+    {
         if (matchDate.isBefore(this.createdAt)) {
             throw RuntimeException("매치 날짜는 생성 날짜 이후여야 합니다") // todo: custom exception
         }
     }
 
     private fun validateContent(
-        content: String
-    ) {
-        if (content.length > 1000) {
+        content:String
+    )
+    {
+        if(content.length >1000){
             throw RuntimeException("내용이 너무 깁니다.") //todo : custom exception
         }
     }
