+++ src/main/kotlin/com/teamsparta/tikitaka/domain/match/service/v1/MatchServiceImpl.kt
@@ -6,6 +6,8 @@ import com.teamsparta.tikitaka.domain.match.dto.PostMatchRequest
 import com.teamsparta.tikitaka.domain.match.dto.UpdateMatchRequest
 import com.teamsparta.tikitaka.domain.match.model.Match
 import com.teamsparta.tikitaka.domain.match.repository.MatchRepository
+import org.springframework.data.domain.Page
+import org.springframework.data.domain.Pageable
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
@@ -15,23 +17,22 @@ import org.springframework.transaction.annotation.Transactional
 class MatchServiceImpl(
     private val matchRepository: MatchRepository,
 
-    ) : MatchService {
+):MatchService {
 
     @Transactional
     override fun postMatch(
         request: PostMatchRequest
-    ): MatchStatusResponse {
+    ): MatchStatusResponse
+    {
 
-        matchRepository.save(
-            Match.of(
-                title = request.title,
-                matchDate = request.matchDate,
-                location = request.location,
-                content = request.content,
-                matchStatus = false,
-                teamId = request.teamId,
-            )
-        )
+        matchRepository.save(Match.of(
+            title = request.title,
+            matchDate = request.matchDate,
+            location = request.location,
+            content = request.content,
+            matchStatus = false,
+            teamId = request.teamId,
+        ))
         //todo : team 구인공고 상태 변경
 
         return MatchStatusResponse.from()
@@ -40,9 +41,10 @@ class MatchServiceImpl(
     @Transactional
     override fun updateMatch(
         matchId: Long, request: UpdateMatchRequest
-    ): MatchStatusResponse {
+    ): MatchStatusResponse
+    {
         matchRepository.findByIdOrNull(matchId)
-            ?.let { it.updateMatch(request) }
+            ?. let { it.updateMatch(request) }
             ?: throw RuntimeException("") //todo : custom exception
 
         return MatchStatusResponse.from()
@@ -51,21 +53,23 @@ class MatchServiceImpl(
     @Transactional
     override fun deleteMatch(
         matchId: Long
-    ): MatchStatusResponse {
+    ): MatchStatusResponse
+    {
         matchRepository.findByIdOrNull(matchId)
-            ?.let { it.softDelete() }
+            ?.let {it.softDelete()}
             ?: throw RuntimeException("Match not found") //todo : custom exception
         return MatchStatusResponse.from()
     }
 
-    override fun getMatches(): List<MatchResponse> {
-        return matchRepository.findByDeletedAtIsNull()
+    override fun getMatches(pageable: Pageable): Page<MatchResponse> {
+        return matchRepository.findByDeletedAtIsNull(pageable)
             .map { match -> MatchResponse.from(match) }
     }
 
     override fun getMatchDetails(
         matchId: Long
-    ): MatchResponse {
+    ): MatchResponse
+    {
         return matchRepository.findByIdOrNull(matchId)
             ?.let { match -> MatchResponse.from(match) }
             ?: throw RuntimeException("Match not found") //todo : custom exception
